requirements.txt
python-etcd
salt==2014.1
python-dateutil==2.2
gevent>=1.0
greenlet>=0.3.2
pytz